COMP: Simplify ITKVideoBridgeOpenCV dependency
When turning on the CMake flag to build the BridgeOpenCV module in
Video for the Slicer branch of ITK, compilation was failing on MacOSX.
Comparison against the successfully building ITK trunk led to finding
a patch that was applied[1] to fix the dependency and missing library
exports. This patch results in the Slicer ITK compiling with this
module enabled. This patch is a prerequisite to adding the
OpenCV extension module to Slicer.

[1] https://github.com/InsightSoftwareConsortium/ITK/commit/412dd83a680cc024c12f971ec57f76881c7eab78